Including the 32 and 64 bit DLLs
