+++ .gitignore
@@ -1,0 +1,3 @@
+*.txt
+!LICENSE.txt
+
